fix: move attempt data tests from student view to api tests
